+++ src/app/(members)/community/[id]/clientPage.tsx
@@ -1,5 +1,5 @@
 import Video from "@/components/ui/video";
-import { communityVideos } from "@/lib/globals";
+import { communityVideos } from "@/lib/mock";
 
 export function ClientPage({ videoID }: { videoID: string }) {
   const communityVideo = communityVideos.filter(
