+++ src/lib/mock.ts
@@ -1,0 +1,86 @@
+// Mock data for videos
+export const userVideos = [
+  {
+    id: "1",
+    title: "Session 1: Introduction to Mindfulness",
+    thumbnail: "/thumbnail.png?height=180&width=320",
+    url: "https://www.youtube.com/embed/z_P4O3yrrms?si=Rcy1estHwXILhz8m",
+    duration: "52:40",
+    date: "Feb 3th, 2025",
+  },
+  {
+    id: "2",
+    title: "Session 2: Stress Management Techniques",
+    thumbnail: "/thumbnail.png?height=180&width=320",
+    url: "https://www.youtube.com/embed/z_P4O3yrrms?si=Rcy1estHwXILhz8m",
+    duration: "52:40",
+    date: "Feb 9th, 2025",
+  },
+  {
+    id: "3",
+    title: "Session 3: Building Healthy Relationships",
+    thumbnail: "/thumbnail.png?height=180&width=320",
+    url: "https://www.youtube.com/embed/z_P4O3yrrms?si=Rcy1estHwXILhz8m",
+    duration: "52:40",
+    date: "Feb 16th, 2025",
+  },
+  {
+    id: "4",
+    title: "Session 4: Overcoming Anxiety",
+    thumbnail: "/thumbnail.png?height=180&width=320",
+    url: "https://www.youtube.com/embed/z_P4O3yrrms?si=Rcy1estHwXILhz8m",
+    duration: "52:40",
+    date: "Feb 17th, 2025",
+  },
+  {
+    id: "5",
+    title: "Session 5: Positive Thinking Strategies",
+    thumbnail: "/thumbnail.png?height=180&width=320",
+    url: "https://www.youtube.com/embed/z_P4O3yrrms?si=Rcy1estHwXILhz8m",
+    duration: "52:40",
+    date: "Feb 25th, 2025",
+  },
+  {
+    id: "6",
+    title: "Session 6: Goal Setting and Achievement",
+    duration: "52:40",
+    thumbnail: "/thumbnail.png?height=180&width=320",
+    url: "https://www.youtube.com/embed/z_P4O3yrrms?si=Rcy1estHwXILhz8m",
+    date: "Feb 30th, 2025",
+  },
+];
+
+export const communityVideos = [
+  {
+    id: "c1",
+    title: "Meditation Basics",
+    duration: "15:30",
+    thumbnail: "/thumbnail.png?height=180&width=320",
+    url: "https://www.youtube.com/embed/z_P4O3yrrms?si=Rcy1estHwXILhz8m",
+    date: "Jan 3th, 2025",
+  },
+  {
+    id: "c2",
+    title: "Stress Relief Techniques",
+    duration: "22:45",
+    thumbnail: "/thumbnail.png?height=180&width=320",
+    url: "https://www.youtube.com/embed/z_P4O3yrrms?si=Rcy1estHwXILhz8m",
+    date: "Jan 15th, 2025",
+  },
+  {
+    id: "c3",
+    title: "Mindfulness Practice",
+    duration: "18:20",
+    thumbnail: "/thumbnail.png?height=180&width=320",
+    url: "https://www.youtube.com/embed/z_P4O3yrrms?si=Rcy1estHwXILhz8m",
+    date: "Jan 27th, 2025",
+  },
+  {
+    id: "c4",
+    title: "Guided Relaxation",
+    duration: "25:10",
+    thumbnail: "/thumbnail.png?height=180&width=320",
+    url: "https://www.youtube.com/embed/z_P4O3yrrms?si=Rcy1estHwXILhz8m",
+    date: "Feb 3rd, 2025",
+  },
+];
